Создана новая директория impl со всеми вытекающими; изменены названия всез сервисов на соответствующие; добавлено 3 метода с конца в adminController; бины теперь подключаются через конструктор

+++ RETURN/src/main/java/com/example/RETURN/repositories/UserRepository.java
@@ -5,17 +5,16 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
+
     boolean existsByEmail(String email);
 
     boolean existsByUserName(String userName);
 
     Optional<User> findByUserName(String userName);
 
-    Optional<User> findUserById(long id);
+    Optional<User> findUserById(int id);
 
     boolean existsOrderByUserName(String name);
-
-
 
 }
